+++ CHANGELOG.md
@@ -48,6 +48,7 @@ v5.10.0
 
 **CI**
 - Add PROCESSORS property to tests to enforce correct parallelism (!2445)
+- Allow multiple tests per `.tst` file (!2509)
 
 **AcousticSolver**
 - Use native c++ random number generator instead of boost (!2379)